osd: Obtain write lock before attempting to clean objects in      a target PG when requesting reservations. If we fail to      obtain the lock, fail the reservation request and retry      once the object is unlocked.

+++ src/osd/PrimaryLogPG.cc
@@ -1742,6 +1742,13 @@ void PrimaryLogPG::do_pg_op(OpRequestRef op)
           return;
         }
 
+        // Delete any leftover objects from previous failed migrations
+        pg_t source_pg = get_source_pg_from_hash(start_obj);
+        if (!pool_migration_target_delete(source_pg, start_obj, op)) {
+          result = -EBUSY;
+          break;
+        }
+
         // Store the op until local reservations are complete and we can reply to the source PG
         pending_pool_migration_reservation_ops.push_back(op);
 
@@ -15794,10 +15801,10 @@ bool PrimaryLogPG::pool_migration_source_delete(hobject_t oid)
   ceph_assert(obc);
   OpContextUPtr ctx = simple_opc_create(obc);
 
-  if (!ctx->lock_manager.get_pool_migration_write(oid, obc)) {
+  if (!ctx->lock_manager.get_pool_migration_write(oid, obc, std::nullopt)) {
     close_op_ctx(ctx.release());
     // Lock acquisition failed - object is already in pending set
-    dout(20) << "pool migration delayed on " << oid
+    dout(20) << __func__ << " pool migration delayed on " << oid
              << "; could not get lock, will retry" << dendl;
     return false;
   }
@@ -15848,7 +15855,9 @@ bool PrimaryLogPG::pool_migration_source_delete(hobject_t oid)
   return true;
 }
 
-void PrimaryLogPG::pool_migration_target_delete(const pg_t &source_pg, const hobject_t &watermark)
+bool PrimaryLogPG::pool_migration_target_delete(const pg_t &source_pg,
+                                                const hobject_t &watermark,
+                                                OpRequestRef& op)
 {
   dout(20) << __func__ << " deleting objects with watermark >= " << watermark << dendl;
 
@@ -15915,8 +15924,15 @@ void PrimaryLogPG::pool_migration_target_delete(const pg_t &source_pg, const hob
       }
 
       OpContextUPtr ctx = simple_opc_create(obc);
-      ctx->at_version = get_next_version();
 
+      if (!ctx->lock_manager.get_pool_migration_write(obj, obc, op)) {
+        dout(20) << __func__ << " could not get lock on " << obj
+                 << ", op will be retried" << dendl;
+        close_op_ctx(ctx.release());
+        return false;
+      }
+
+      ctx->at_version = get_next_version();
       int ret = _delete_oid(ctx.get(), true, false, true);
       ceph_assert(ret == 0);
       if (obc->obs.oi.is_omap()) {
@@ -15935,6 +15951,7 @@ void PrimaryLogPG::pool_migration_target_delete(const pg_t &source_pg, const hob
   }
 
   dout(20) << __func__ << " deleted " << deleted << " stale objects" << dendl;
+  return true;
 }
 
 void PrimaryLogPG::handle_pool_migration_quiesce_complete()
